Редактирует тесты
