+++ src/components/question-artist/question-artist.e2e.test.js
@@ -2,31 +2,11 @@ import React from 'react';
 import {configure, shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
+import {questionArtist} from '../../utils/questions-test.js';
+
 import QuestionArtist from './question-artist.jsx';
 
 configure({adapter: new Adapter()});
-
-const question = {
-  type: `artist`,
-  answers: [
-    {
-      artist: `Паша`,
-      picture: `https://focus.ua/storage/pub/images/2017/2615387.jpg`
-    },
-    {
-      artist: `Лена`,
-      picture: `https://focus.ua/storage/pub/images/2017/2615387.jpg`
-    },
-    {
-      artist: `Жора`,
-      picture: `https://focus.ua/storage/pub/images/2017/2615387.jpg`
-    },
-  ],
-  song: {
-    artist: `Лена`,
-    src: `https://music.yandex.ru/album/10798364/track/66306611`
-  }
-};
 
 describe(`QuestionArtistComponent`, () => {
   it(`Change input value`, () => {
@@ -38,7 +18,7 @@ describe(`QuestionArtistComponent`, () => {
 
     const screen = shallow(
         <QuestionArtist
-          question={question}
+          question={questionArtist}
           onAnswer={onAnswer}
         />
     );
@@ -50,7 +30,7 @@ describe(`QuestionArtistComponent`, () => {
 
     expect(onAnswer).toHaveBeenCalledTimes(1);
 
-    expect(onAnswer.mock.calls[0][0]).toMatchObject(question);
+    expect(onAnswer.mock.calls[0][0]).toMatchObject(questionArtist);
     expect(onAnswer.mock.calls[0][1]).toMatchObject(userAnswer);
   });
 });
