+++ src/components/question-genre/question-genre.test.js
@@ -1,35 +1,14 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import QuestionGenre from './question-genre'
 
-const question = {
-  type: `genre`,
-  genre: `rep`,
-  answers: [
-    {
-      genre: `rep`,
-      src: `https://music.yandex.ru/album/10798364/track/66306611`
-    },
-    {
-      genre: `rock`,
-      src: `https://music.yandex.ru/album/10798364/track/66306611`
-    },
-    {
-      genre: `hip-hop`,
-      src: `https://music.yandex.ru/album/10798364/track/66306611`
-    },
-    {
-      genre: `pop`,
-      src: `https://music.yandex.ru/album/10798364/track/66306611`
-    }
-  ]
-};
+import {questionGenre} from '../../utils/questions-test.js';
 
+import QuestionGenre from './question-genre';
 
 describe(`QuestionArtistComponent`, () => {
   it(`QuestionArtistSnapshot`, () => {
     const tree = renderer.create(<QuestionGenre
-      question={question}
+      question={questionGenre}
       onAnswer={() => {}}
     />).toJSON();
